修改int64 》》 uint64

+++ example/testringbuffer.go
@@ -10,7 +10,7 @@ import (
 
 func main() {
 	rbuffer := ringbuffer.RingBuffer{}
-	rbuffer.RingBufferInit(int64(3))
+	rbuffer.RingBufferInit(3)
 	fmt.Println(rbuffer.GetCurrentReadIndex())
 	fmt.Println(rbuffer.GetCurrentWriteIndex())
 	/*bytes := make([]byte, 20)
